Add detail-surf.php with functionality to show Surf destination details
Created a new file named detail-surf.php that displays details of the Surf destination. When the Surf image is clicked from the Destinations page, it navigates to this file and shows relevant content about Surf activities, packages, and accommodations. Also added a back button linking back to destination.php for easy navigation.